requirements.txt
python-dotenv==0.20.0
slack-bolt==1.13.1
slack-sdk==3.16.0
